requirements.txt
basyx-python-sdk>=0.2.2
toml~=0.10.2
pandas~=1.5.3